Add image cache to improve responsiveness.
The idea is that some feeds use the same images for certain things, and
the image cache intends to speed that up.  Eventually, an image prefetch
scheme will be implemented, that will fetch images, in a thread, from
the next few feed items, as a way of reducing the lag between a feed
item is selected and its images appear.  These prefetched images will be
stored in the image cache.